Optimized stand-alone k-means clustering
Weaviate currently uses k-means clustering for product quantization and
the implementation of k-means clustering was tied to this use case.
This patch factors the PQ encode/decode logic into a separate struct and
rewrites k-means clustering to improve performance while adding
k-means++ initialization.

The running time of the new implementation is roughly comparable to the
previous version, but there a some major improvements:

    1) The total number of heap allocations is drastically reduced
    (previously every iteration of k-means used at least one allocation
    per data point).

    2) The total number of bytes allocated is reduced by ~90%.

    3) With k-means++ initialization we produce much better clusterings
    compared to random initialization, at least for some data sets.
    For the same number of iterations we see a 50% reduction in the
    within-cluster sum of squares (the k-means objective function) when
    comparing to random initialization.